requirements.txt
Flask==3.0.0
Pillow==11.2.1
qrcode==8.2
gunicorn==21.2.0